config: Cast price_stop_level to (float)

+++ config/EURUSD_M1.h
@@ -23,7 +23,7 @@ struct Stg_DeMarker_Params_M1 : StgParams {
     signal_close_method = -1;
     signal_close_level = (float)0.5;
     price_stop_method = 0;
-    price_stop_level = 0.0;
+    price_stop_level = (float)0.0;
     tick_filter_method = 1;
     max_spread = 0;
   }
